fix: client explorer test stubs and transport labels
- Deduplicate Armeria client PSI stubs via shared test support
- Remove unrelated route.explorer bundle keys from clients PR
- Omit WebClient transport label for Retrofit SessionProtocol+EndpointGroup
